[Kay Smarczewski]: Biblatex-Support: recognize addbibresource
Patch taken from
http://sourceforge.net/tracker/?func=detail&aid=3315923&group_id=52322&atid=466458

Copied description:
-----
When using biblatex the bib file is included by \addbibresource{library.bib}
Vim-Latex does not recognize this command and so the outline does not
work.

The attached tiny patch replaces the search functions in the texviewer.
It is a quite simple patch, just works for this special case and makes
latex suite recognize the bib file.

The diff is against the actual git master branch.