Improved indentation (multiple braces on one line)
Now, something like:

\ifthenelse{...}{
	Yes.
}{
	No.
}

is properly indented (as above). The whole line is indented according
to the position in the line which causes the least indentation (after
as many braces as possible where closed).

It does not work if environments start/end on the same line. This
could, however, be handled by similar code, but I think this should
not be TeXed in this way, anyway.