Enhance the errorformat by track overfull \hbox
There was some issue with 'overfull \hbox' messages. They also output
some garbage from the input file. If that contains parenthesis, the
tracing of filenames is messed up.

It is recommended, that 'overfull \hbox' messages are not ignored.
Otherwise, the scanning of the log file may fail completely, due to
a bug(?) in vim, see vim/vim#1126